Changed data type of large constants MILLIS_IN_A_DAY and MILLIS_IN_A_MINUTE from int to long Fixed pemdas (added parentheses) to the stop/start alarm computation of weekly date which incorrectly set the weekly alarm as only 7 days after the set date and not how many days it really is from the current date.
Ex: Set alarm is Jan 11, 5pm and date today is Mar 7
Alarm will be set at Jan 17, X hours (7 days ) when it should be set
56 days from Jan 11, 5pm which is Mar 7, 5pm.

+++ app/src/main/java/wearable/userwatch/Constants.java
@@ -91,7 +91,7 @@ public class Constants {
 
     //TIME DATE CONSTANTS
     public static final int DAYS_IN_A_WEEK = 7;
-    public static final int MILLIS_IN_A_DAY = 24 * 60 * 60 * 1000;
-    public static final int MILLIS_IN_A_MINUTE = 60 * 1000;
+    public static final long MILLIS_IN_A_DAY = 24 * 60 * 60 * 1000;
+    public static final long MILLIS_IN_A_MINUTE = 60 * 1000;
 
 }
